Replace symbol hacks with AssociatedTypeRequirementsVisitor hacks
I built a generic version of the hacks I debugged with @a2 for decomposition.

This one does not use any of the symbol hacks from before, but instead is betting on the implementations on changing the default associated type of their parent protocol.

This means we can get rid of all the linux workarounds 🎉